+++ packages/lib/src/lua/serde/compress_decompress.rs
@@ -1,6 +1,8 @@
 use async_compression::tokio::write::{
     BrotliDecoder, BrotliEncoder, GzipDecoder, GzipEncoder, ZlibDecoder, ZlibEncoder,
 };
+use blocking::unblock;
+use lz4_flex::{compress_prepend_size, decompress_size_prepended};
 use mlua::prelude::*;
 use tokio::io::AsyncWriteExt;
 
@@ -8,27 +10,44 @@ use tokio::io::AsyncWriteExt;
 pub enum CompressDecompressFormat {
     Brotli,
     GZip,
+    LZ4,
     ZLib,
 }
 
 #[allow(dead_code)]
 impl CompressDecompressFormat {
     pub fn detect_from_bytes(bytes: impl AsRef<[u8]>) -> Option<Self> {
-        let bytes = bytes.as_ref();
-        if bytes[0..4] == [0x0B, 0x24, 0x72, 0x68] {
-            Some(Self::Brotli)
-        } else if bytes[0..3] == [0x1F, 0x8B, 0x08] {
-            Some(Self::GZip)
-        }
-        // https://stackoverflow.com/a/54915442
-        else if (bytes[0..2] == [0x78, 0x01])
-            || (bytes[0..2] == [0x78, 0x5E])
-            || (bytes[0..2] == [0x78, 0x9C])
-            || (bytes[0..2] == [0x78, 0xDA])
-        {
-            Some(Self::ZLib)
-        } else {
-            None
+        match bytes.as_ref() {
+            // https://github.com/PSeitz/lz4_flex/blob/main/src/frame/header.rs#L28
+            b if b.len() >= 4
+                && matches!(
+                    u32::from_le_bytes(b[0..4].try_into().unwrap()),
+                    0x184D2204 | 0x184C2102
+                ) =>
+            {
+                Some(Self::LZ4)
+            }
+            // https://github.com/dropbox/rust-brotli/blob/master/src/enc/brotli_bit_stream.rs#L2805
+            b if b.len() >= 4
+                && matches!(
+                    b[0..3],
+                    [0xE1, 0x97, 0x81] | [0xE1, 0x97, 0x82] | [0xE1, 0x97, 0x80]
+                ) =>
+            {
+                Some(Self::Brotli)
+            }
+            // https://github.com/rust-lang/flate2-rs/blob/main/src/gz/mod.rs#L135
+            b if b.len() >= 3 && matches!(b[0..3], [0x1F, 0x8B, 0x08]) => Some(Self::GZip),
+            // https://stackoverflow.com/a/43170354
+            b if b.len() >= 2
+                && matches!(
+                    b[0..2],
+                    [0x78, 0x01] | [0x78, 0x5E] | [0x78, 0x9C] | [0x78, 0xDA]
+                ) =>
+            {
+                Some(Self::ZLib)
+            }
+            _ => None,
         }
     }
 
@@ -49,12 +68,13 @@ impl<'lua> FromLua<'lua> for CompressDecompressFormat {
             match s.to_string_lossy().to_ascii_lowercase().trim() {
                 "brotli" => Ok(Self::Brotli),
                 "gzip" => Ok(Self::GZip),
+                "lz4" => Ok(Self::LZ4),
                 "zlib" => Ok(Self::ZLib),
                 kind => Err(LuaError::FromLuaConversionError {
                     from: value.type_name(),
                     to: "CompressDecompressFormat",
                     message: Some(format!(
-                        "Invalid format '{kind}', valid formats are:  brotli, gzip, zlib"
+                        "Invalid format '{kind}', valid formats are:  brotli, gzip, lz4, zlib"
                     )),
                 }),
             }
@@ -89,6 +109,10 @@ pub async fn compress<'lua>(
                 .write_all(source.as_ref())
                 .await?
         }
+        CompressDecompressFormat::LZ4 => {
+            let source = source.as_ref().to_vec();
+            bytes = unblock(move || compress_prepend_size(&source)).await;
+        }
     }
     Ok(bytes)
 }
@@ -113,6 +137,12 @@ pub async fn decompress<'lua>(
             ZlibDecoder::new(&mut bytes)
                 .write_all(source.as_ref())
                 .await?
+        }
+        CompressDecompressFormat::LZ4 => {
+            let source = source.as_ref().to_vec();
+            bytes = unblock(move || decompress_size_prepended(&source))
+                .await
+                .map_err(LuaError::external)?;
         }
     }
     Ok(bytes)